2021/3/22 22:41

+++ api/src/main/java/org/leafbook/api/respAbs/indexPage/BrowseHistoryAbs.java
@@ -9,6 +9,6 @@ public class BrowseHistoryAbs {
     private Long topicId;
     private String topicTitle;
     private String topicDesc;
-    private Integer likeNumber;
+    private Long likeNumber;
     private List<EntryData> entryDataList;
 }
